updated file, folder and dependency names and install.json to razorpay
