+++ catalog/model/payment/razorpay.php
@@ -1,5 +1,5 @@
 <?php
-namespace Opencart\Catalog\Model\Extension\OcPaymentExample\Payment;
+namespace Opencart\Catalog\Model\Extension\Razorpay\Payment;
 class Razorpay extends \Opencart\System\Engine\Model {
 	public function getMethods(array $address): array {
 		// echo(json_encode($this->load->language('extension/oc_payment_example/payment/razorpay')));
