+++ install.json
@@ -1,5 +1,6 @@
 {
-  "name": "OpenCart Payment Example",
+  "name": "Razorpay",
+  "code": "razorpay",
   "version": "1.0",
   "author": "OpenCart Ltd",
   "link": "https://www.opencart.com"
